pyproject.toml
[tool.poetry]
name = "flask_cognito_lib"
version = "0.1.0"
description = " A Flask extension that supports protecting routes with AWS Cognito following OAuth 2.0 best practices"
license = "MIT"
authors = ["mblackgeo <18327836+mblackgeo@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/mblackgeo/flask-cognito-lib"
repository = "https://github.com/mblackgeo/flask-cognito-lib"
keywords = ["Flask", "Extension", "OAuth", "Cognito"]
classifiers = [
    "Framework :: Flask",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
    { include = "flask_cognito_lib", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
requests = "^2.27.1"
python-jose = {extras = ["crytography"], version = "^3.3.0"}
Flask-JWT-Extended = "^4.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.941"
black = "^22.1.0"
flake8 = "^4.0.1"
typed-ast = "^1.5.2"
pylint = "^2.12.2"
types-requests = "^2.27.13"
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["flask_coginito_lib"]